rename SourceDefinition to DatasourceProvider; move all to central file
