--- libs/datamodel/core/src/configuration/source/builtin/mod.rs
@@ -1,10 +1,0 @@
-mod mysql_source_definition;
-mod postgres_source_definition;
-pub mod shared_validation;
-mod simple_source;
-mod sqlite_source_definition;
-
-pub use mysql_source_definition::*;
-pub use postgres_source_definition::*;
-pub use simple_source::*;
-pub use sqlite_source_definition::*;
