--- libs/datamodel/core/src/configuration/source/builtin/shared_validation.rs
@@ -1,12 +1,0 @@
-use crate::StringFromEnvVar;
-
-pub fn validate_url(name: &str, expected_protocol: &str, url: StringFromEnvVar) -> Result<StringFromEnvVar, String> {
-    if url.value.starts_with(expected_protocol) {
-        Ok(url)
-    } else {
-        Err(format!(
-            "The URL for datasource `{}` must start with the protocol `{}`.",
-            name, expected_protocol
-        ))
-    }
-}
