+++ libs/datamodel/core/src/configuration/source/mod.rs
@@ -1,12 +1,15 @@
+mod builtin_datasource_providers;
 mod loader;
 mod serializer;
+mod simple_source;
 mod traits;
 
-pub mod builtin;
+//pub mod builtin;
 
 // TODO: i think these constants should move to a more central place.
-pub use builtin::{MYSQL_SOURCE_NAME, POSTGRES_SOURCE_NAME, SQLITE_SOURCE_NAME};
-//pub use json::{render_sources_to_json, render_sources_to_json_value, sources_from_json_value_with_plugins};
+//pub use builtin::{MYSQL_SOURCE_NAME, POSTGRES_SOURCE_NAME, SQLITE_SOURCE_NAME};
+pub use builtin_datasource_providers::*;
 pub use loader::*;
 pub use serializer::*;
+pub use simple_source::*;
 pub use traits::*;
