+++ libs/datamodel/core/src/configuration/source/traits.rs
@@ -1,13 +1,9 @@
 use crate::StringFromEnvVar;
 use datamodel_connector::Connector;
 
-// TODO: Probably rename everything. Terminology here is messy.
-
-/// Trait for custom sources.
-///
 /// A source is basically the datamodel equivalent of a connector.
 pub trait Source: Send + Sync {
-    /// Gets the name of the implementing connector.
+    /// Gets the name of the provider.
     fn connector_type(&self) -> &str;
 
     /// Gets the name of the source configuration block.
@@ -24,11 +20,11 @@ pub trait Source: Send + Sync {
     fn connector(&self) -> &Box<dyn Connector>;
 }
 
-pub trait SourceDefinition {
+pub trait DatasourceProvider {
     /// Passes the provider arg from the datasource. Must return true for all provider names it can handle.
     fn is_provider(&self, provider: &str) -> bool;
 
-    /// Instantiates a new source, using the given name, url and detailed arguments.
+    /// Instantiates a new source
     fn create(
         &self,
         name: &str,
